+++ shell/run.docker.sh
@@ -1,9 +1,10 @@
 #!/usr/bin/env bash
 
-#@echo off
-
 cd .. || exit 1
 
-# Add `BUILDKIT_PROGRESS=plain` before the docker-compose command
-# to log output of the build process.
+# To log output of the build process.
+export BUILDKIT_PROGRESS=plain
+
+export APP_PROFILES=docker,headless
+
 docker-compose -p automate-idea-to-social up -d